added more initializers and stuff for tests
