+++ Purchases/Caching/RCInMemoryCachedObject.m
@@ -8,6 +8,7 @@
 #import <Foundation/Foundation.h>
 
 #import "RCInMemoryCachedObject.h"
+#import "RCInMemoryCachedObject+Protected.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -17,6 +18,8 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, nullable) NSDate *lastUpdatedAt;
 @property (nonatomic, assign) int cacheDurationInSeconds;
 @property (nonatomic, nullable) id cachedInstance;
+
+@property (nonatomic, nullable) NSDate *stubbedNow;
 
 @end
 
@@ -30,9 +33,18 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (instancetype)initWithCacheDurationInSeconds:(int)cacheDurationInSeconds
                                  lastUpdatedAt:(nullable NSDate *)lastUpdatedAt {
+    return [self initWithCacheDurationInSeconds:cacheDurationInSeconds
+                                  lastUpdatedAt:lastUpdatedAt
+                                     stubbedNow:nil];
+}
+
+- (instancetype)initWithCacheDurationInSeconds:(int)cacheDurationInSeconds
+                                 lastUpdatedAt:(nullable NSDate *)lastUpdatedAt
+                                    stubbedNow:(nullable NSDate *)stubbedNow {
     if (self == [super init]) {
         self.cacheDurationInSeconds = cacheDurationInSeconds;
         self.lastUpdatedAt = lastUpdatedAt;
+        self.stubbedNow = stubbedNow;
     }
     return self;
 }
@@ -42,7 +54,7 @@ NS_ASSUME_NONNULL_BEGIN
         return YES;
     }
 
-    NSTimeInterval timeSinceLastCheck = -1 * [self.lastUpdatedAt timeIntervalSinceNow];
+    NSTimeInterval timeSinceLastCheck = -1 * [self.lastUpdatedAt timeIntervalSinceDate:self.now];
     return timeSinceLastCheck >= self.cacheDurationInSeconds;
 }
 
@@ -66,6 +78,14 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (void)updateCacheTimestampWithDate:(NSDate *)date {
     self.lastUpdatedAt = date;
+}
+
+- (NSDate *)now {
+    if (self.stubbedNow) {
+        return self.stubbedNow;
+    } else {
+        return [NSDate date];
+    }
 }
 
 @end
