+++ PurchasesTests/DeviceCacheTests.swift
@@ -14,49 +14,6 @@ class DeviceCacheTests: XCTestCase {
 
     private var deviceCache: RCDeviceCache! = nil
 
-    class MockUserDefaults: UserDefaults {
-
-        var stringForKeyCalledValue: String? = nil
-        var setObjectForKeyCalledValue: String? = nil
-        var removeObjectForKeyCalledValues: Array<String> = []
-        var dataForKeyCalledValue: String? = nil
-        var objectForKeyCalledValue: String? = nil
-        var setBoolForKeyCalledValue: String? = nil
-        var setValueForKeyCalledValue: String? = nil
-
-        var mockValues: [String: Any] = [:]
-
-        override func string(forKey defaultName: String) -> String? {
-            stringForKeyCalledValue = defaultName
-            return mockValues[defaultName] as? String
-        }
-
-        override func removeObject(forKey defaultName: String) {
-            removeObjectForKeyCalledValues.append(defaultName)
-            mockValues.removeValue(forKey: defaultName)
-        }
-
-        override func set(_ value: Any?, forKey defaultName: String) {
-            setObjectForKeyCalledValue = defaultName
-            mockValues[defaultName] = value
-        }
-
-        override func data(forKey defaultName: String) -> Data? {
-            dataForKeyCalledValue = defaultName
-            return mockValues[defaultName] as? Data
-        }
-
-        override func object(forKey defaultName: String) -> Any? {
-            objectForKeyCalledValue = defaultName
-            return mockValues[defaultName]
-        }
-
-        override func set(_ value: Bool, forKey defaultName: String) {
-            setValueForKeyCalledValue = defaultName
-            mockValues[defaultName] = value
-        }
-    }
-
     override func setUp() {
         self.mockUserDefaults = MockUserDefaults()
         self.deviceCache = RCDeviceCache(mockUserDefaults)
@@ -115,14 +72,26 @@ class DeviceCacheTests: XCTestCase {
         expect(self.deviceCache.isPurchaserInfoCacheStale()).to(beTrue())
     }
 
-    func testCacheIsStaleIfLongerThanFiveMinutes() {
+    func testPurchaserInfoCacheIsStaleIfLongerThanFiveMinutes() {
         let oldDate: Date! = Calendar.current.date(byAdding: .minute, value: -(6), to: Date())
         self.deviceCache = RCDeviceCache(mockUserDefaults, stubbedNow: oldDate)
         self.deviceCache.cachePurchaserInfo(Data(), forAppUserID: "waldo")
-        
+
         expect(self.deviceCache.isPurchaserInfoCacheStale()).to(beFalse())
         self.deviceCache.stubbedNow = Date()
         expect(self.deviceCache.isPurchaserInfoCacheStale()).to(beTrue())
+    }
+
+    func testOfferingsCacheIsStaleIfLongerThanFiveMinutes() {
+        let oldDate: Date! = Calendar.current.date(byAdding: .minute, value: -(6), to: Date())
+        self.deviceCache = RCDeviceCache(mockUserDefaults, stubbedNow: oldDate)
+        let offerings = Purchases.Offerings()
+        self.deviceCache.cacheOfferings(offerings)
+
+        expect(self.deviceCache.isOfferingsCacheStale()).to(beFalse())
+        // wip, this needs to be initialized with a mock InMemoryCachedObject
+        self.deviceCache.stubbedNow = Date()
+        expect(self.deviceCache.isOfferingsCacheStale()).to(beTrue())
     }
 
     func testPurchaserInfoIsProperlyCached() {
@@ -132,5 +101,4 @@ class DeviceCacheTests: XCTestCase {
         expect(self.deviceCache.cachedPurchaserInfoData(forAppUserID: "cesar")).to(equal(data))
         expect(self.mockUserDefaults.setObjectForKeyCalledValue).to(equal("com.revenuecat.userdefaults.purchaserInfo.cesar"))
     }
-
 }
