+++ PurchasesTests/PurchasesTests-Bridging-Header.h
@@ -24,4 +24,5 @@
 #include <Purchases/RCEntitlementInfos.h>
 #include <Purchases/RCDeviceCache.h>
 #include <Purchases/RCDeviceCache+Protected.h>
+#include <Purchases/RCInMemoryCachedObject+Protected.h>
 #include <Purchases/RCIdentityManager.h>
